fix: update `duration` parameter to support double precision in animateCamera methods

+++ android/src/main/java/com/google/android/react/navsdk/NavAutoModule.java
@@ -623,7 +623,7 @@ public class NavAutoModule extends NativeNavAutoModuleSpec
           }
 
           Map<String, Object> map = cameraPosition.toHashMap();
-          map.put("duration", duration != null ? duration.intValue() : 0);
+          map.put("duration", duration != null ? duration.doubleValue() : 0.0);
           mMapViewController.animateCamera(map);
           promise.resolve(null);
         });
